`cargo clippy --fix`

+++ src/asciibird/display.rs
@@ -28,7 +28,7 @@ impl BaseDisplay for Display {
     fn update_cursor(&mut self, game_objects: &[Box<dyn GameObject>]) {
         if let Some(player_object) = take_game_object::<PlayerCharacter>(game_objects) {
             let x = player_object.coordinate.x;
-            let abs_diff = x.abs_diff(self.view_cursor().x as usize);
+            let abs_diff = x.abs_diff(self.view_cursor().x);
             if abs_diff > 1 && abs_diff < (self.screen_height() as usize - 2_usize) {
                 return;
             }
